fixed buttons migration to new flutter recommendation :)
