+++ lib/ui/designsystem/components/ct_buttons.dart
@@ -1,15 +1,56 @@
+
+import 'package:core_theming/ui/designsystem/utils/scale.dart' as scale;
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
 import 'package:core_theming/ui/designsystem/theme.dart';
 
-/// This creates a default buttons that can be usd all across the app
+/// This is a button builder that will be used to build the
+/// various properties we need for the buttons that will be created.
+/// The
+class _NewButtonBuilder {
+  final double edgeInset17 = scale.value(17.0);
+  final VoidCallback? onPressed;
+  final String label;
+  final Color? btnPrimaryColor;
+  final Color? btnTextColor;
+  final OutlinedBorder btnShape;
 
-class CTButtonGrey extends StatelessWidget {
+  /* final Color? color;*/
+
+  _NewButtonBuilder({
+    required this.onPressed,
+    required this.label,
+    required this.btnPrimaryColor,
+    required this.btnTextColor,
+    required this.btnShape,
+
+
+  });
+
+  ElevatedButton build(){
+    return ElevatedButton(
+      child: Text(this.label),
+      onPressed: this.onPressed == null ? null : this.onPressed,
+      style: ElevatedButton.styleFrom(
+        padding: EdgeInsets.all(edgeInset17),
+        primary: this.btnPrimaryColor,
+        onPrimary: this.btnTextColor,
+        elevation: 5,
+        shape: this.btnShape,
+      ),
+
+    );
+  }
+}
+
+
+class CTButtonDefault extends StatelessWidget {
+  final double radiusCircular10 = scale.value(10.0);
   final String label;
   final VoidCallback? onPressed;
 
-  CTButtonGrey({
+  CTButtonDefault ({
     required this.label,
     this.onPressed,
   });
@@ -20,99 +61,219 @@ class CTButtonGrey extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final themeData = CTTheme.of(context);
-    return _ButtonBuilder(
-      color: themeData.gray,
-      textColor: themeData.black,
-      splashColor: themeData.deepGray,
+    return _NewButtonBuilder(
       label: this.label,
+      btnPrimaryColor: themeData.deepGray,
+      btnTextColor: themeData.black,
+      btnShape: RoundedRectangleBorder(
+        borderRadius: new BorderRadius.circular(radiusCircular10),
+      ),
       onPressed: this.onPressed,
     ).build();
   }
 }
-
-class CTButtonBlack extends StatelessWidget {
+class CTButtonRedWhite extends StatelessWidget {
+  final double radiusCircular10 = scale.value(10.0);
   final String label;
   final VoidCallback? onPressed;
 
-  CTButtonBlack({
+  CTButtonRedWhite ({
     required this.label,
     this.onPressed,
   });
 
-  /// We build our buttons calling the JFthemeData and the buttonBuilder
+  /// We build our buttons calling the CTThemeData and the buttonBuilder
   /// class.
   /// This can be repeated for all the buttons we want to build
   @override
   Widget build(BuildContext context) {
     final themeData = CTTheme.of(context);
-    return _ButtonBuilder(
-      color: themeData.black,
-      textColor: themeData.white,
-      splashColor: themeData.white,
+    return _NewButtonBuilder(
       label: this.label,
+      btnPrimaryColor: themeData.reddish,
+      btnTextColor: CTTheme.of(context).white,
+      btnShape: RoundedRectangleBorder(
+        borderRadius: new BorderRadius.circular(radiusCircular10),
+      ),
       onPressed: this.onPressed,
     ).build();
   }
 }
-
-class CTButtonRed extends StatelessWidget {
+class CTButtonRedBlack extends StatelessWidget {
+  final double radiusCircular10 = scale.value(10.0);
   final String label;
   final VoidCallback? onPressed;
 
-  CTButtonRed({
+  CTButtonRedBlack ({
     required this.label,
     this.onPressed,
   });
 
-  /// We build our buttons calling the JFthemeData and the buttonBuilder
+  /// We build our buttons calling the CTThemeData and the buttonBuilder
   /// class.
   /// This can be repeated for all the buttons we want to build
   @override
   Widget build(BuildContext context) {
     final themeData = CTTheme.of(context);
-    return _ButtonBuilder(
-      color: themeData.warmRed,
-      textColor: themeData.white,
-      splashColor: themeData.white,
+    return _NewButtonBuilder(
       label: this.label,
+      btnPrimaryColor: themeData.reddish,
+      btnTextColor: CTTheme.of(context).black,
+      btnShape: RoundedRectangleBorder(
+        borderRadius: new BorderRadius.circular(radiusCircular10),
+      ),
+      onPressed: this.onPressed,
+    ).build();
+  }
+}
+class CTButtonWarmRedWhite extends StatelessWidget {
+  final double radiusCircular10 = scale.value(10.0);
+  final String label;
+  final VoidCallback? onPressed;
+
+  CTButtonWarmRedWhite ({
+    required this.label,
+    this.onPressed,
+  });
+
+  /// We build our buttons calling the CTThemeData and the buttonBuilder
+  /// class.
+  /// This can be repeated for all the buttons we want to build
+  @override
+  Widget build(BuildContext context) {
+    final themeData = CTTheme.of(context);
+    return _NewButtonBuilder(
+      label: this.label,
+      btnPrimaryColor: themeData.warmRed,
+      btnTextColor: CTTheme.of(context).white,
+      btnShape: RoundedRectangleBorder(
+        borderRadius: new BorderRadius.circular(radiusCircular10),
+      ),
+      onPressed: this.onPressed,
+    ).build();
+  }
+}
+class CTButtonWarmRedBlack extends StatelessWidget {
+  final double radiusCircular10 = scale.value(10.0);
+  final String label;
+  final VoidCallback? onPressed;
+
+  CTButtonWarmRedBlack ({
+    required this.label,
+    this.onPressed,
+  });
+
+  /// We build our buttons calling the CTThemeData and the buttonBuilder
+  /// class.
+  /// This can be repeated for all the buttons we want to build
+  @override
+  Widget build(BuildContext context) {
+    final themeData = CTTheme.of(context);
+    return _NewButtonBuilder(
+      label: this.label,
+      btnPrimaryColor: themeData.warmRed,
+      btnTextColor: CTTheme.of(context).black,
+      btnShape: RoundedRectangleBorder(
+        borderRadius: new BorderRadius.circular(radiusCircular10),
+      ),
+      onPressed: this.onPressed,
+    ).build();
+  }
+}
+class CTButtonBlackWhite extends StatelessWidget {
+  final double radiusCircular10 = scale.value(10.0);
+  final String label;
+  final VoidCallback? onPressed;
+
+  CTButtonBlackWhite ({
+    required this.label,
+    this.onPressed,
+  });
+
+  /// We build our buttons calling the CTThemeData and the buttonBuilder
+  /// class.
+  /// This can be repeated for all the buttons we want to build
+  @override
+  Widget build(BuildContext context) {
+    final themeData = CTTheme.of(context);
+    return _NewButtonBuilder(
+      label: this.label,
+      btnPrimaryColor: themeData.black,
+      btnTextColor: CTTheme.of(context).white,
+      btnShape: RoundedRectangleBorder(
+        borderRadius: new BorderRadius.circular(radiusCircular10),
+      ),
+      onPressed: this.onPressed,
+    ).build();
+  }
+}
+class CTButtonLimeGreenBlack extends StatelessWidget {
+  final double radiusCircular10 = scale.value(10.0);
+  final String label;
+  final VoidCallback? onPressed;
+
+  CTButtonLimeGreenBlack ({
+    required this.label,
+    this.onPressed,
+  });
+
+  /// We build our buttons calling the CTThemeData and the buttonBuilder
+  /// class.
+  /// This can be repeated for all the buttons we want to build
+  @override
+  Widget build(BuildContext context) {
+    final themeData = CTTheme.of(context);
+    return _NewButtonBuilder(
+      label: this.label,
+      btnPrimaryColor: themeData.limeGreenish,
+      btnTextColor: CTTheme.of(context).black,
+      btnShape: RoundedRectangleBorder(
+        borderRadius: new BorderRadius.circular(radiusCircular10),
+      ),
+      onPressed: this.onPressed,
+    ).build();
+  }
+}
+class CTButtonLimeGreenWhite extends StatelessWidget {
+  final double radiusCircular10 = scale.value(10.0);
+  final String label;
+  final VoidCallback? onPressed;
+
+  CTButtonLimeGreenWhite ({
+    required this.label,
+    this.onPressed,
+  });
+
+  /// We build our buttons calling the CTThemeData and the buttonBuilder
+  /// class.
+  /// This can be repeated for all the buttons we want to build
+  @override
+  Widget build(BuildContext context) {
+    final themeData = CTTheme.of(context);
+    return _NewButtonBuilder(
+      label: this.label,
+      btnPrimaryColor: themeData.limeGreenish,
+      btnTextColor: CTTheme.of(context).white,
+      btnShape: RoundedRectangleBorder(
+        borderRadius: new BorderRadius.circular(radiusCircular10),
+      ),
       onPressed: this.onPressed,
     ).build();
   }
 }
 
-/// This is a button builder that will be used to build the
-/// various properties we need for the buttons that will be created.
-class _ButtonBuilder {
-  final Color? color;
-  final Color? textColor;
-  final Color? splashColor;
-  final double elevation;
-  final String label;
-  final VoidCallback? onPressed;
-  final double borderRadius;
-  final double buttonWidth;
+/*
+You can also increase size by wrapping the button
+with a constrained box like:
+child: ConstrainedBox(
+            constraints: BoxConstraints.tightFor(
+            width: 300, height: 200
+            ),
+            child: ElevatedButton(
+              child: Text('300 x 200'), under the buttons place
+              onPressed: () {},
+            ),
+          ),
 
-  _ButtonBuilder(
-      {required this.color,
-      required this.textColor,
-      required this.splashColor,
-      required this.label,
-      required this.onPressed,
-      this.buttonWidth = double.infinity,
-      this.elevation = 3.0,
-      this.borderRadius = 13.0});
-
-  RaisedButton build() {
-    return RaisedButton(
-
-        onPressed: this.onPressed == null ? null : this.onPressed,
-        color: this.color,
-        textColor: this.textColor,
-        splashColor: this.splashColor,
-        elevation: this.elevation,
-        padding: EdgeInsets.only(top: 17.0, bottom: 17.0),
-        shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(borderRadius)),
-        child: Text(this.label));
-  }
-}
+          see an implementation in the main.dart
+*/
